Checkpointer and BgWriter code closer to PG 9.2.
Rename checkpoint.c to checkpointer.c. And move the code from bgwriter.c to
checkpointer.c and also renames most of corresponding data structures to refect
the clear ownership and association. This commit brings it as close as possible
to PostgreSQL 9.2.

Reference to PostgreSQL related commits:
commit 806a2aee3791244bf0f916729bfdb5489936e068
    Split work of bgwriter between 2 processes: bgwriter and checkpointer.
commit bf405ba8e460051e715d0a91442b579e590328ce
    Add new file for checkpointer.c
commit 8f28789bffd03e6d865039aff00c68db54d9f574
    Rename BgWriterShmem/Request to CheckpointerShmem/Request
commit d843589e5ab361dd4738dab5c9016e704faf4153
    Fix management of pendingOpsTable in auxiliary processes.
