+++ src/backend/postmaster/Makefile
@@ -13,7 +13,7 @@ top_builddir = ../../..
 include $(top_builddir)/src/Makefile.global
 override CPPFLAGS := -I$(top_srcdir)/src/backend/gp_libpq_fe $(CPPFLAGS)
 
-OBJS = autovacuum.o bgwriter.o checkpoint.o fork_process.o seqserver.o pgarch.o pgstat.o \
+OBJS = autovacuum.o bgwriter.o checkpointer.o fork_process.o seqserver.o pgarch.o pgstat.o \
 	postmaster.o primary_mirror_mode.o primary_mirror_transition_client.o syslogger.o \
 	perfmon.o backoff.o perfmon_segmentinfo.o \
 	sendalert.o alertseverity.o autostats.o walwriter.o
