--- src/backend/postmaster/checkpoint.c
@@ -1,821 +1,0 @@
-/*-------------------------------------------------------------------------
- *
- * checkpoint.c
- *
- * The checkpoint server process handles periodically doing checkpoints and regular checkpoint
- * requests.  Previously, the background writer use to do this duty.  But we ended up with
- * a deadlock with the new FileRep functionality, so this functionality was split out into its
- * own server.
- *
- * It will automatically dispatch a checkpoint after a certain amount of time has
- * elapsed since the last one, and it can be signaled to perform requested
- * checkpoints as well.  (The GUC parameter that mandates a checkpoint every
- * so many WAL segments is implemented by having backends signal the checkpoint server
- * when they fill WAL segments; the checkpoint server itself doesn't watch for the
- * condition.)
- *
- * The checkpoint server is started by the postmaster as soon as the startup subprocess
- * finishes.  It remains alive until the postmaster commands it to terminate.
- * Normal termination is by SIGUSR2, which instructs the bgwriter to execute
- * a shutdown checkpoint and then exit(0).	(All backends must be stopped
- * before SIGUSR2 is issued!)  Emergency termination is by SIGQUIT; like any
- * backend, the checkpoint server will simply abort and exit on SIGQUIT.
- *
- * If the checkpoint server exits unexpectedly, the postmaster treats that the same
- * as a backend crash: shared memory may be corrupted, so remaining backends
- * should be killed by SIGQUIT and then a recovery cycle started.  (Even if
- * shared memory isn't corrupted, we have lost information about which
- * files need to be fsync'd for the next checkpoint, and so a system
- * restart needs to be forced.)
- *
- *
- * Portions Copyright (c) 1996-2008, PostgreSQL Global Development Group
- *
- *-------------------------------------------------------------------------
- */
-#include "postgres.h"
-
-#include <signal.h>
-#include <time.h>
-#include <unistd.h>
-
-#include "access/xlog_internal.h"
-#include "libpq/pqsignal.h"
-#include "miscadmin.h"
-#include "pgstat.h"
-#include "postmaster/bgwriter.h"
-#include "storage/fd.h"
-#include "storage/freespace.h"
-#include "storage/ipc.h"
-#include "storage/lwlock.h"
-#include "storage/pmsignal.h"
-#include "storage/shmem.h"
-#include "storage/smgr.h"
-#include "storage/spin.h"
-#include "tcop/tcopprot.h"
-#include "utils/guc.h"
-#include "utils/memutils.h"
-#include "utils/resowner.h"
-#include "postmaster/fork_process.h"
-#include "postmaster/postmaster.h"
-
-#include "tcop/tcopprot.h" /* quickdie() */
-
-/*----------
- * Shared memory area for communication between checkpoint server and backends
- *
- * The ckpt counters allow backends to watch for completion of a checkpoint
- * request they send.  Here's how it works:
- *	* At start of a checkpoint, bgwriter reads (and clears) the request flags
- *	  and increments ckpt_started, while holding ckpt_lck.
- *	* On completion of a checkpoint, bgwriter sets ckpt_done to
- *	  equal ckpt_started.
- *	* On failure of a checkpoint, bgwriter increments ckpt_failed
- *	  and sets ckpt_done to equal ckpt_started.
- *
- * The algorithm for backends is:
- *	1. Record current values of ckpt_failed and ckpt_started, and
- *	   set request flags, while holding ckpt_lck.
- *	2. Send signal to request checkpoint.
- *	3. Sleep until ckpt_started changes.  Now you know a checkpoint has
- *	   begun since you started this algorithm (although *not* that it was
- *	   specifically initiated by your signal), and that it is using your flags.
- *	4. Record new value of ckpt_started.
- *	5. Sleep until ckpt_done >= saved value of ckpt_started.  (Use modulo
- *	   arithmetic here in case counters wrap around.)  Now you know a
- *	   checkpoint has started and completed, but not whether it was
- *	   successful.
- *	6. If ckpt_failed is different from the originally saved value,
- *	   assume request failed; otherwise it was definitely successful.
- *
- * ckpt_flags holds the OR of the checkpoint request flags sent by all
- * requesting backends since the last checkpoint start.  The flags are
- * chosen so that OR'ing is the correct way to combine multiple requests.
- *----------
- */
-
-typedef struct
-{
-	pid_t		checkpoint_server_pid;	/* PID of checkpoint server (0 if not started) */
-
-	slock_t		ckpt_lck;		/* protects all the ckpt_* fields */
-
-	int			ckpt_started;	/* advances when checkpoint starts */
-	int			ckpt_done;		/* advances when checkpoint done */
-	int			ckpt_failed;	/* advances when checkpoint fails */
-
-	int			ckpt_flags;		/* checkpoint flags, as defined in xlog.h */
-} CheckpointShmemStruct;
-
-static CheckpointShmemStruct *CheckpointShmem;
-
-/* interval for calling AbsorbFsyncRequests in CheckpointWriteDelay */
-#define WRITES_PER_ABSORB		1000
-
-/*
- * GUC parameters
- */
-int			CheckPointTimeout = 300;
-int			CheckPointWarning = 30;
-double		CheckPointCompletionTarget = 0.5;
-
-/*
- * Flags set by interrupt handlers for later service in the main loop.
- */
-static volatile sig_atomic_t got_SIGHUP = false;
-static volatile sig_atomic_t checkpoint_requested = false;
-static volatile sig_atomic_t shutdown_requested = false;
-
-/*
- * Private state
- */
-static bool am_checkpoint_server = false;
-
-static bool ckpt_active = false;
-
-/* these values are valid when ckpt_active is true: */
-static time_t ckpt_start_time;
-static XLogRecPtr ckpt_start_recptr;
-static double ckpt_cached_elapsed;
-
-static time_t last_checkpoint_time;
-
-/* Prototypes for private functions */
-
-static bool IsCheckpointOnSchedule(double progress);
-static bool ImmediateCheckpointRequested(void);
-
-/* Signal handlers */
-
-static void CheckpointSigHupHandler(SIGNAL_ARGS);
-static void ReqCheckpointHandler(SIGNAL_ARGS);
-static void ReqShutdownHandler(SIGNAL_ARGS);
-
-
-
-/*
- * Main entry point for bgwriter process
- *
- * This is invoked from AuxiliaryProcessMain, which has already created the basic
- * execution environment, but not enabled signals yet.
- */
-void
-CheckpointMain(void)
-{
-	sigjmp_buf	local_sigjmp_buf;
-	MemoryContext checkpoint_context;
-
-	Assert(CheckpointShmem != NULL);
-	CheckpointShmem->checkpoint_server_pid = MyProcPid;
-	am_checkpoint_server = true;
-
-	/*
-	 * If possible, make this process a group leader, so that the postmaster
-	 * can signal any child processes too.  (bgwriter probably never has
-	 * any child processes, but for consistency we make all postmaster
-	 * child processes do this.)
-	 */
-#ifdef HAVE_SETSID
-	if (setsid() < 0)
-		elog(FATAL, "setsid() failed: %m");
-#endif
-
-	/*
-	 * Properly accept or ignore signals the postmaster might send us
-	 *
-	 * Note: we deliberately ignore SIGTERM, because during a standard Unix
-	 * system shutdown cycle, init will SIGTERM all processes at once.	We
-	 * want to wait for the backends to exit, whereupon the postmaster will
-	 * tell us it's okay to shut down (via SIGUSR2).
-	 *
-	 * SIGUSR1 is presently unused; keep it spare in case someday we want this
-	 * process to participate in sinval messaging.
-	 */
-	pqsignal(SIGHUP, CheckpointSigHupHandler);	/* set flag to read config file */
-	pqsignal(SIGINT, ReqCheckpointHandler);		/* request checkpoint */
-	pqsignal(SIGTERM, SIG_IGN); /* ignore SIGTERM */
-	pqsignal(SIGQUIT, quickdie);		/* hard crash time: nothing bg-writer specific, just use the standard */
-	pqsignal(SIGALRM, SIG_IGN);
-	pqsignal(SIGPIPE, SIG_IGN);
-	pqsignal(SIGUSR1, SIG_IGN); /* reserve for sinval */
-	pqsignal(SIGUSR2, ReqShutdownHandler);		/* request shutdown */
-
-	/*
-	 * Reset some signals that are accepted by postmaster but not here
-	 */
-	pqsignal(SIGCHLD, SIG_DFL);
-	pqsignal(SIGTTIN, SIG_DFL);
-	pqsignal(SIGTTOU, SIG_DFL);
-	pqsignal(SIGCONT, SIG_DFL);
-	pqsignal(SIGWINCH, SIG_DFL);
-
-	/* We allow SIGQUIT (quickdie) at all times */
-#ifdef HAVE_SIGPROCMASK
-	sigdelset(&BlockSig, SIGQUIT);
-#else
-	BlockSig &= ~(sigmask(SIGQUIT));
-#endif
-
-	/*
-	 * Initialize so that first time-driven event happens at the correct time.
-	 */
-	last_checkpoint_time = time(NULL);
-
-	/*
-	 * Create a resource owner to keep track of our resources (currently only
-	 * buffer pins).
-	 */
-	CurrentResourceOwner = ResourceOwnerCreate(NULL, "Checkpoint Server");
-
-	/*
-	 * Create a memory context that we will do all our work in.  We do this so
-	 * that we can reset the context during error recovery and thereby avoid
-	 * possible memory leaks.  Formerly this code just ran in
-	 * TopMemoryContext, but resetting that would be a really bad idea.
-	 */
-	checkpoint_context = AllocSetContextCreate(TopMemoryContext,
-											 "Background Writer",
-											 ALLOCSET_DEFAULT_MINSIZE,
-											 ALLOCSET_DEFAULT_INITSIZE,
-											 ALLOCSET_DEFAULT_MAXSIZE);
-	MemoryContextSwitchTo(checkpoint_context);
-
-	/*
-	 * If an exception is encountered, processing resumes here.
-	 *
-	 * See notes in postgres.c about the design of this coding.
-	 */
-	if (sigsetjmp(local_sigjmp_buf, 1) != 0)
-	{
-		/* Since not using PG_TRY, must reset error stack by hand */
-		error_context_stack = NULL;
-
-		/* Prevent interrupts while cleaning up */
-		HOLD_INTERRUPTS();
-
-		/* Report the error to the server log */
-		EmitErrorReport();
-
-		/*
-		 * These operations are really just a minimal subset of
-		 * AbortTransaction().	We don't have very many resources to worry
-		 * about in bgwriter, but we do have LWLocks, buffers, and temp files.
-		 */
-		LWLockReleaseAll();
-		AbortBufferIO();
-		UnlockBuffers();
-		/* buffer pins are released here: */
-		ResourceOwnerRelease(CurrentResourceOwner,
-							 RESOURCE_RELEASE_BEFORE_LOCKS,
-							 false, true);
-		/* we needn't bother with the other ResourceOwnerRelease phases */
-		AtEOXact_Buffers(false);
-		AtEOXact_Files();
-		AtEOXact_HashTables(false);
-
-		/* Warn any waiting backends that the checkpoint failed. */
-		if (ckpt_active)
-		{
-			/* use volatile pointer to prevent code rearrangement */
-			volatile CheckpointShmemStruct *css = CheckpointShmem;
-
-			SpinLockAcquire(&css->ckpt_lck);
-			css->ckpt_failed++;
-			css->ckpt_done = css->ckpt_started;
-			SpinLockRelease(&css->ckpt_lck);
-		}
-
-		/*
-		 * Now return to normal top-level context and clear ErrorContext for
-		 * next time.
-		 */
-		MemoryContextSwitchTo(checkpoint_context);
-		FlushErrorState();
-
-		/* Flush any leaked data in the top-level context */
-		MemoryContextResetAndDeleteChildren(checkpoint_context);
-
-		/* Now we can allow interrupts again */
-		RESUME_INTERRUPTS();
-
-		/*
-		 * Sleep at least 1 second after any error.  A write error is likely
-		 * to be repeated, and we don't want to be filling the error logs as
-		 * fast as we can.
-		 */
-		pg_usleep(1000000L);
-
-		/*
-		 * Close all open files after any error.  This is helpful on Windows,
-		 * where holding deleted files open causes various strange errors.
-		 * It's not clear we need it elsewhere, but shouldn't hurt.
-		 */
-		smgrcloseall();
-	}
-
-	/* We can now handle ereport(ERROR) */
-	PG_exception_stack = &local_sigjmp_buf;
-
-	/*
-	 * Unblock signals (they were blocked when the postmaster forked us)
-	 */
-	PG_SETMASK(&UnBlockSig);
-
-	/*
-	 * Loop forever
-	 */
-	for (;;)
-	{
-		bool		do_checkpoint = false;
-		int			flags = 0;
-		time_t		now;
-		int			elapsed_secs;
-
-		/*
-		 * Emergency bailout if postmaster has died.  This is to avoid the
-		 * necessity for manual cleanup of all postmaster children.
-		 */
-		if (!PostmasterIsAlive(true))
-			exit(1);
-
-		/*
-		 * Process any requests or signals received recently.
-		 */
-		AbsorbFsyncRequests();
-
-		if (got_SIGHUP)
-		{
-			got_SIGHUP = false;
-			ProcessConfigFile(PGC_SIGHUP);
-		}
-		if (checkpoint_requested)
-		{
-			checkpoint_requested = false;
-			do_checkpoint = true;
-			BgWriterStats.m_requested_checkpoints++;
-		}
-		if (shutdown_requested)
-		{
-			/*
-			 * From here on, elog(ERROR) should end with exit(1), not send
-			 * control back to the sigsetjmp block above
-			 */
-			ExitOnAnyError = true;
-			/* Close down the database */
-			ShutdownXLOG(0, 0);
-
-			/* Normal exit from the checkpoint server is here */
-			if (Debug_print_server_processes ||
-				Debug_print_qd_mirroring ||
-				log_min_messages <= DEBUG5)
-			{
-				ereport((log_min_messages > DEBUG5 ? LOG : DEBUG5),
-						(errmsg("normal exit from checkpoint server")));
-			}
-			proc_exit(0);		/* done */
-		}
-
-		/*
-		 * Force a checkpoint if too much time has elapsed since the last one.
-		 * Note that we count a timed checkpoint in stats only when this
-		 * occurs without an external request, but we set the CAUSE_TIME flag
-		 * bit even if there is also an external request.
-		 */
-		now = time(NULL);
-		elapsed_secs = now - last_checkpoint_time;
-		if (elapsed_secs >= CheckPointTimeout)
-		{
-			if (!do_checkpoint)
-				BgWriterStats.m_timed_checkpoints++;
-			do_checkpoint = true;
-			flags |= CHECKPOINT_CAUSE_TIME;
-		}
-
-		/*
-		 * Do a checkpoint if requested, otherwise do one cycle of
-		 * dirty-buffer writing.
-		 */
-		if (do_checkpoint)
-		{
-			/* use volatile pointer to prevent code rearrangement */
-			volatile CheckpointShmemStruct *css = CheckpointShmem;
-
-			/*
-			 * Atomically fetch the request flags to figure out what kind of a
-			 * checkpoint we should perform, and increase the started-counter
-			 * to acknowledge that we've started a new checkpoint.
-			 */
-			SpinLockAcquire(&css->ckpt_lck);
-			flags |= css->ckpt_flags;
-			css->ckpt_flags = 0;
-			css->ckpt_started++;
-			SpinLockRelease(&css->ckpt_lck);
-
-			/*
-			 * We will warn if (a) too soon since last checkpoint (whatever
-			 * caused it) and (b) somebody set the CHECKPOINT_CAUSE_XLOG flag
-			 * since the last checkpoint start.  Note in particular that this
-			 * implementation will not generate warnings caused by
-			 * CheckPointTimeout < CheckPointWarning.
-			 */
-			if ((flags & CHECKPOINT_CAUSE_XLOG) &&
-				elapsed_secs < CheckPointWarning)
-				ereport(LOG,
-						(errmsg("checkpoints are occurring too frequently (%d seconds apart)",
-								elapsed_secs),
-						 errhint("Consider increasing the configuration parameter \"checkpoint_segments\".")));
-
-			/*
-			 * Initialize bgwriter-private variables used during checkpoint.
-			 */
-			ckpt_active = true;
-			ckpt_start_recptr = GetInsertRecPtr();
-			ckpt_start_time = now;
-			ckpt_cached_elapsed = 0;
-
-			/*
-			 * Do the checkpoint.
-			 */
-			CreateCheckPoint(flags);
-
-			/*
-			 * After any checkpoint, close all smgr files.	This is so we
-			 * won't hang onto smgr references to deleted files indefinitely.
-			 */
-			smgrcloseall();
-
-			/*
-			 * Signal the background writer to also call smgrcloseall.
-			 */
-			RequestCheckpointSmgrCloseAll();
-
-			/*
-			 * Indicate checkpoint completion to any waiting backends.
-			 */
-			SpinLockAcquire(&css->ckpt_lck);
-			css->ckpt_done = css->ckpt_started;
-			SpinLockRelease(&css->ckpt_lck);
-
-			ckpt_active = false;
-
-			/*
-			 * Note we record the checkpoint start time not end time as
-			 * last_checkpoint_time.  This is so that time-driven checkpoints
-			 * happen at a predictable spacing.
-			 */
-			last_checkpoint_time = now;
-		}
-
-		/* Nap for one second. */
-		if (!(got_SIGHUP || checkpoint_requested || shutdown_requested))
-			pg_usleep(1000000L);
-	}
-}
-
-/*
- * Returns true if an immediate checkpoint request is pending.	(Note that
- * this does not check the *current* checkpoint's IMMEDIATE flag, but whether
- * there is one pending behind it.)
- */
-static bool
-ImmediateCheckpointRequested(void)
-{
-	if (checkpoint_requested)
-	{
-		volatile CheckpointShmemStruct *css = CheckpointShmem;
-
-		/*
-		 * We don't need to acquire the ckpt_lck in this case because we're
-		 * only looking at a single flag bit.
-		 */
-		if (css->ckpt_flags & CHECKPOINT_IMMEDIATE)
-			return true;
-	}
-	return false;
-}
-
-/*
- * CheckpointWriteDelay -- yield control to bgwriter during a checkpoint
- *
- * This function is called after each page write performed by BufferSync().
- * It is responsible for keeping the bgwriter's normal activities in
- * progress during a long checkpoint, and for throttling BufferSync()'s
- * write rate to hit checkpoint_completion_target.
- *
- * The checkpoint request flags should be passed in; currently the only one
- * examined is CHECKPOINT_IMMEDIATE, which disables delays between writes.
- *
- * 'progress' is an estimate of how much of the work has been done, as a
- * fraction between 0.0 meaning none, and 1.0 meaning all done.
- */
-void
-CheckpointWriteDelay(int flags, double progress)
-{
-	static int	absorb_counter = WRITES_PER_ABSORB;
-
-	/* Do nothing if checkpoint is being executed by non-checkpointer process */
-	if (!am_checkpoint_server)
-		return;
-
-	/*
-	 * Take a nap, unless we're behind
-	 * schedule, in which case we just try to catch up as quickly as possible.
-	 */
-	if (!(flags & CHECKPOINT_IMMEDIATE) &&
-		!shutdown_requested &&
-		!ImmediateCheckpointRequested() &&
-		IsCheckpointOnSchedule(progress))
-	{
-		if (got_SIGHUP)
-		{
-			got_SIGHUP = false;
-			ProcessConfigFile(PGC_SIGHUP);
-		}
-
-		AbsorbFsyncRequests();
-		absorb_counter = WRITES_PER_ABSORB;
-
-		/*
-		 * Report interim activity statistics to the stats collector.
-		 */
-		pgstat_send_bgwriter();
-
-		/*
-		 * This sleep used to be connected to bgwriter_delay, typically 200ms.
-		 * That resulted in more frequent wakeups if not much work to do.
-		 * Checkpointer and bgwriter are no longer related so take the Big
-		 * Sleep.
-		 */
-		pg_usleep(100000L);
-	}
-	else if (--absorb_counter <= 0)
-	{
-		/*
-		 * Absorb pending fsync requests after each WRITES_PER_ABSORB write
-		 * operations even when we don't sleep, to prevent overflow of the
-		 * fsync request queue.
-		 */
-		AbsorbFsyncRequests();
-		absorb_counter = WRITES_PER_ABSORB;
-	}
-}
-
-/*
- * IsCheckpointOnSchedule -- are we on schedule to finish this checkpoint
- *		 in time?
- *
- * Compares the current progress against the time/segments elapsed since last
- * checkpoint, and returns true if the progress we've made this far is greater
- * than the elapsed time/segments.
- */
-static bool
-IsCheckpointOnSchedule(double progress)
-{
-	XLogRecPtr	recptr;
-	struct timeval now;
-	double		elapsed_xlogs,
-				elapsed_time;
-
-	Assert(ckpt_active);
-
-	/* Scale progress according to checkpoint_completion_target. */
-	progress *= CheckPointCompletionTarget;
-
-	/*
-	 * Check against the cached value first. Only do the more expensive
-	 * calculations once we reach the target previously calculated. Since
-	 * neither time or WAL insert pointer moves backwards, a freshly
-	 * calculated value can only be greater than or equal to the cached value.
-	 */
-	if (progress < ckpt_cached_elapsed)
-		return false;
-
-	/*
-	 * Check progress against WAL segments written and checkpoint_segments.
-	 *
-	 * We compare the current WAL insert location against the location
-	 * computed before calling CreateCheckPoint. The code in XLogInsert that
-	 * actually triggers a checkpoint when checkpoint_segments is exceeded
-	 * compares against RedoRecptr, so this is not completely accurate.
-	 * However, it's good enough for our purposes, we're only calculating an
-	 * estimate anyway.
-	 */
-	recptr = GetInsertRecPtr();
-	elapsed_xlogs =
-		(((double) (int32) (recptr.xlogid - ckpt_start_recptr.xlogid)) * XLogSegsPerFile +
-		 ((double) recptr.xrecoff - (double) ckpt_start_recptr.xrecoff) / XLogSegSize) /
-		CheckPointSegments;
-
-	if (progress < elapsed_xlogs)
-	{
-		ckpt_cached_elapsed = elapsed_xlogs;
-		return false;
-	}
-
-	/*
-	 * Check progress against time elapsed and checkpoint_timeout.
-	 */
-	gettimeofday(&now, NULL);
-	elapsed_time = ((double) (now.tv_sec - ckpt_start_time) +
-					now.tv_usec / 1000000.0) / CheckPointTimeout;
-
-	if (progress < elapsed_time)
-	{
-		ckpt_cached_elapsed = elapsed_time;
-		return false;
-	}
-
-	/* It looks like we're on schedule. */
-	return true;
-}
-
-
-/* --------------------------------
- *		signal handler routines
- * --------------------------------
- */
-
-/* SIGHUP: set flag to re-read config file at next convenient time */
-static void
-CheckpointSigHupHandler(SIGNAL_ARGS)
-{
-	got_SIGHUP = true;
-}
-
-/* SIGINT: set flag to run a normal checkpoint right away */
-static void
-ReqCheckpointHandler(SIGNAL_ARGS)
-{
-	checkpoint_requested = true;
-}
-
-/* SIGUSR2: set flag to run a shutdown checkpoint and exit */
-static void
-ReqShutdownHandler(SIGNAL_ARGS)
-{
-	shutdown_requested = true;
-}
-
-
-/* --------------------------------
- *		communication with backends
- * --------------------------------
- */
-
-/*
- * CheckpointShmemSize
- *		Compute space needed for checkpoint server related shared memory
- */
-Size
-CheckpointShmemSize(void)
-{
-	Size		size;
-
-	size = sizeof(CheckpointShmemStruct);
-
-	return size;
-}
-
-/*
- * CheckpointShmemInit
- *		Allocate and initialize checkpoint server related shared memory
- */
-void
-CheckpointShmemInit(void)
-{
-	bool		found;
-
-	CheckpointShmem = (CheckpointShmemStruct *)
-		ShmemInitStruct("Checkpoint Server Data",
-						CheckpointShmemSize(),
-						&found);
-	if (CheckpointShmem == NULL)
-		ereport(FATAL,
-				(errcode(ERRCODE_OUT_OF_MEMORY),
-				 errmsg("not enough shared memory for checkpoint server")));
-	if (found)
-		return;					/* already initialized */
-
-	MemSet(CheckpointShmem, 0, sizeof(CheckpointShmemStruct));
-	SpinLockInit(&CheckpointShmem->ckpt_lck);
-}
-
-/*
- * RequestCheckpoint
- *		Called in backend processes to request a checkpoint
- *
- * flags is a bitwise OR of the following:
- *	CHECKPOINT_IS_SHUTDOWN: checkpoint is for database shutdown.
- *	CHECKPOINT_IMMEDIATE: finish the checkpoint ASAP,
- *		ignoring checkpoint_completion_target parameter.
- *	CHECKPOINT_FORCE: force a checkpoint even if no XLOG activity has occured
- *		since the last one (implied by CHECKPOINT_IS_SHUTDOWN).
- *	CHECKPOINT_WAIT: wait for completion before returning (otherwise,
- *		just signal bgwriter to do it, and return).
- *	CHECKPOINT_CAUSE_XLOG: checkpoint is requested due to xlog filling.
- *		(This affects logging, and in particular enables CheckPointWarning.)
- */
-void
-RequestCheckpoint(int flags)
-{
-	/* use volatile pointer to prevent code rearrangement */
-	volatile CheckpointShmemStruct *css = CheckpointShmem;
-	int			old_failed,
-				old_started;
-
-	/*
-	 * If in a standalone backend, just do it ourselves.
-	 */
-	if (!IsPostmasterEnvironment)
-	{
-		/*
-		 * There's no point in doing slow checkpoints in a standalone backend,
-		 * because there's no other backends the checkpoint could disrupt.
-		 */
-		CreateCheckPoint(flags | CHECKPOINT_IMMEDIATE);
-
-		/*
-		 * After any checkpoint, close all smgr files.	This is so we won't
-		 * hang onto smgr references to deleted files indefinitely.
-		 */
-		smgrcloseall();
-
-		return;
-	}
-
-	/*
-	 * Atomically set the request flags, and take a snapshot of the counters.
-	 * When we see ckpt_started > old_started, we know the flags we set here
-	 * have been seen by bgwriter.
-	 *
-	 * Note that we OR the flags with any existing flags, to avoid overriding
-	 * a "stronger" request by another backend.  The flag senses must be
-	 * chosen to make this work!
-	 */
-	SpinLockAcquire(&css->ckpt_lck);
-
-	old_failed = css->ckpt_failed;
-	old_started = css->ckpt_started;
-	css->ckpt_flags |= flags;
-
-	SpinLockRelease(&css->ckpt_lck);
-
-	/*
-	 * Send signal to request checkpoint.  When not waiting, we consider
-	 * failure to send the signal to be nonfatal.
-	 */
-	if (CheckpointShmem->checkpoint_server_pid == 0)
-		elog((flags & CHECKPOINT_WAIT) ? ERROR : LOG,
-			 "could not request checkpoint because checkpoint server not running");
-	if (kill(CheckpointShmem->checkpoint_server_pid, SIGINT) != 0)
-		elog((flags & CHECKPOINT_WAIT) ? ERROR : LOG,
-			 "could not signal for checkpoint: %m");
-
-	/*
-	 * If requested, wait for completion.  We detect completion according to
-	 * the algorithm given above.
-	 */
-	if (flags & CHECKPOINT_WAIT)
-	{
-		int			new_started,
-					new_failed;
-
-		/* Wait for a new checkpoint to start. */
-		for (;;)
-		{
-			SpinLockAcquire(&css->ckpt_lck);
-			new_started = css->ckpt_started;
-			SpinLockRelease(&css->ckpt_lck);
-
-			if (new_started != old_started)
-				break;
-
-			CHECK_FOR_INTERRUPTS();
-			pg_usleep(100000L);
-		}
-
-		/*
-		 * We are waiting for ckpt_done >= new_started, in a modulo sense.
-		 */
-		for (;;)
-		{
-			int			new_done;
-
-			SpinLockAcquire(&css->ckpt_lck);
-			new_done = css->ckpt_done;
-			new_failed = css->ckpt_failed;
-			SpinLockRelease(&css->ckpt_lck);
-
-			if (new_done - new_started >= 0)
-				break;
-
-			CHECK_FOR_INTERRUPTS();
-			pg_usleep(100000L);
-		}
-
-		if (new_failed != old_failed)
-			ereport(ERROR,
-					(errmsg("checkpoint request failed"),
-					 errhint("Consult recent messages in the server log for details.")));
-	}
-}
