+++ src/backend/postmaster/test/Makefile
@@ -2,7 +2,7 @@ subdir=src/backend/postmaster
 top_builddir=../../../..
 include $(top_builddir)/src/Makefile.global
 
-TARGETS=syslogger bgwriter
+TARGETS=syslogger checkpointer
 
 include $(top_builddir)/src/backend/mock.mk
-bgwriter.t: $(MOCK_DIR)/backend/storage/lmgr/lwlock_mock.o
+checkpointer.t: $(MOCK_DIR)/backend/storage/lmgr/lwlock_mock.o
