+++ src/backend/postmaster/test/checkpointer_test.c
@@ -4,18 +4,18 @@
 #include "cmockery.h"
 
 #include "c.h"
-#include "../bgwriter.c"
+#include "../checkpointer.c"
 #include "postgres.h"
 
 #define MAX_BGW_REQUESTS 5
 void
 init_request_queue(void)
 {
-	size_t size = sizeof(BgWriterShmemStruct) + sizeof(BgWriterRequest)*MAX_BGW_REQUESTS;
-	BgWriterShmem = (BgWriterShmemStruct *) malloc(size);
-	memset(BgWriterShmem, 0, size);
-	BgWriterShmem->bgwriter_pid = 1234;
-	BgWriterShmem->max_requests = MAX_BGW_REQUESTS;
+	size_t size = sizeof(CheckpointerShmemStruct) + sizeof(CheckpointerRequest)*MAX_BGW_REQUESTS;
+	CheckpointerShmem = (CheckpointerShmemStruct *) malloc(size);
+	memset(CheckpointerShmem, 0, size);
+	CheckpointerShmem->checkpointer_pid = 1234;
+	CheckpointerShmem->max_requests = MAX_BGW_REQUESTS;
 	IsUnderPostmaster = true;
 }
 
@@ -30,33 +30,33 @@ test__ForwardFsyncRequest_enqueue(void **state)
 	int i;
 	RelFileNode dummy = {1,1,1};
 	init_request_queue();
-	expect_value(LWLockAcquire, lockid, BgWriterCommLock);
+	expect_value(LWLockAcquire, lockid, CheckpointerCommLock);
 	expect_value(LWLockAcquire, mode, LW_EXCLUSIVE);
 	will_be_called(LWLockAcquire);
-	expect_value(LWLockRelease, lockid, BgWriterCommLock);
+	expect_value(LWLockRelease, lockid, CheckpointerCommLock);
 	will_be_called(LWLockRelease);
 	/* basic enqueue */
 	ret = ForwardFsyncRequest(dummy, 1);
 	assert_true(ret);
-	assert_true(BgWriterShmem->num_requests == 1);
+	assert_true(CheckpointerShmem->num_requests == 1);
 	/* fill up the queue */
 	for (i=2; i<=MAX_BGW_REQUESTS; i++)
 	{
-		expect_value(LWLockAcquire, lockid, BgWriterCommLock);
+		expect_value(LWLockAcquire, lockid, CheckpointerCommLock);
 		expect_value(LWLockAcquire, mode, LW_EXCLUSIVE);
 		will_be_called(LWLockAcquire);
-		expect_value(LWLockRelease, lockid, BgWriterCommLock);
+		expect_value(LWLockRelease, lockid, CheckpointerCommLock);
 		will_be_called(LWLockRelease);
 		ret = ForwardFsyncRequest(dummy, i);
 		assert_true(ret);
 	}
-	expect_value(LWLockAcquire, lockid, BgWriterCommLock);
+	expect_value(LWLockAcquire, lockid, CheckpointerCommLock);
 	expect_value(LWLockAcquire, mode, LW_EXCLUSIVE);
 	will_be_called(LWLockAcquire);
-	expect_value(LWLockRelease, lockid, BgWriterCommLock);
+	expect_value(LWLockRelease, lockid, CheckpointerCommLock);
 	will_be_called(LWLockRelease);
 #ifdef USE_ASSERT_CHECKING
-	expect_value(LWLockHeldByMe, lockid, BgWriterCommLock);
+	expect_value(LWLockHeldByMe, lockid, CheckpointerCommLock);
 	will_return(LWLockHeldByMe, true);
 #endif
 	/*
@@ -66,8 +66,8 @@ test__ForwardFsyncRequest_enqueue(void **state)
 	 */
 	ret = ForwardFsyncRequest(dummy, 0);
 	assert_false(ret);
-	assert_true(BgWriterShmem->num_requests == BgWriterShmem->max_requests);
-	free(BgWriterShmem);
+	assert_true(CheckpointerShmem->num_requests == CheckpointerShmem->max_requests);
+	free(CheckpointerShmem);
 }
 
 int
