+++ src/include/access/xlog_internal.h
@@ -269,7 +269,7 @@ typedef struct RmgrData
 extern const RmgrData RmgrTable[];
 
 /*
- * Exported to support xlog switching from bgwriter
+ * Exported to support xlog switching from checkpointer
  */
 extern time_t GetLastSegSwitchTime(void);
 extern XLogRecPtr RequestXLogSwitch(void);
