+++ src/include/storage/lwlock.h
@@ -68,7 +68,7 @@ typedef enum LWLockId
 	MultiXactOffsetControlLock,
 	MultiXactMemberControlLock,
 	RelCacheInitLock,
-	BgWriterCommLock,
+	CheckpointerCommLock,
 	TwoPhaseStateLock,
 	TablespaceCreateLock,
 	BtreeVacuumLock,
