+++ src/include/storage/proc.h
@@ -198,10 +198,10 @@ typedef struct PROC_HDR
  * We set aside some extra PGPROC structures for auxiliary processes,
  * ie things that aren't full-fledged backends but need shmem access.
  *
- * Background writer, WAL writer, and autovacuum launcher run during
- * normal operation. Startup process also consumes one slot, but WAL
- * writer and autovacuum launcher are launched only after it has
- * exited (4 slots).
+ * Background writer, checkpointer, WAL writer, and autovacuum launcher run
+ * during normal operation. Startup process also consumes one slot, but WAL
+ * writer and autovacuum launcher are launched only after it has exited (4
+ * slots).
  *
  * FileRep Process uses 
  *			a) 10 slots on Primary
